requirements.txt
absl-py==2.0.0
ago==0.0.95
alembic==1.12.0
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
astunparse==1.6.3
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
blinker==1.6.3
boto3==1.28.62
botocore==1.31.62
bs4==0.0.1
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
cmaes==0.10.0
cmdstanpy==1.2.0
colorlog==6.7.0
constantly==15.1.0
contourpy==1.1.1
cryptography==41.0.4
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.3
datapackage==1.15.2
dateparser==1.1.8
dotmap==1.3.30
elastic-transport==8.4.1
elasticsearch==8.10.0
eod==0.2.1
et-xmlfile==1.1.0
exchange-calendars==4.5
fastapi==0.103.2
faust-cchardet==2.1.19
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.1
frozendict==2.3.8
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
hjson==3.1.0
holidays==0.34
html5lib==1.1
hurry.filesize==0.9
hyperlink==21.0.0
idna==3.4
ijson==3.2.3
importlib-resources==6.1.0
incremental==22.10.0
IProgress==0.4
isodate==0.6.1
itemadapter==0.8.0
itemloaders==1.1.0
jieba3k==0.35.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.14.0
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
langdetect==1.0.9
libclang==16.0.6
linear-tsv==1.1.0
loguru==0.7.2
lxml==4.9.3
Mako==1.2.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
ml-dtypes==0.2.0
mplfinance==0.12.10b0
multitasking==0.0.11
news-please==1.5.35
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.0
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
optuna==3.3.0
optuna-integration==3.3.0
pandas==2.1.1
pandas-datareader==0.10.0
pandas-market-calendars==4.3.1
parsel==1.8.1
patsy==0.5.3
peewee==3.16.3
Pillow==10.0.1
plac==1.4.0
plotly==5.17.0
pmdarima==2.0.3
prophet==1.1.5
Protego==0.3.0
protobuf==4.24.4
psycopg2-binary==2.9.9
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
PyDispatcher==2.0.7
pyluach==2.2.0
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
queuelib==1.6.2
readability-lxml==0.8.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3986==2.0.0
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
s3transfer==0.7.0
scikit-learn==1.3.1
scipy==1.11.3
Scrapy==2.11.0
seaborn==0.13.0
service-identity==23.1.0
sgmllib3k==1.0.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
stanio==0.3.0
starlette==0.27.0
statsmodels==0.14.0
streamlit==1.27.2
ta==0.10.2
tableschema==1.20.2
tabulator==1.53.5
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
tensorflow-metal==1.1.0
termcolor==2.3.0
threadpoolctl==3.2.0
tinysegmenter==0.3
tldextract==3.6.0
toml==0.10.2
toolz==0.12.0
tqdm==4.66.1
Twisted==22.10.0
tzdata==2023.3
tzlocal==5.1
unicodecsv==0.14.1
urllib3==1.26.17
uvicorn==0.23.2
validators==0.22.0
w3lib==2.1.2
warcio==1.7.4
webencodings==0.5.1
Werkzeug==3.0.0
wrapt==1.14.1
xgboost==2.0.0
xlrd==2.0.1
yfinance==0.2.31
zope.interface==6.1
